+++ Scripts/Security/Deploy-ASR.ps1
@@ -87,6 +87,7 @@ https://docs.microsoft.com/en-us/microsoft-365/security/defender-endpoint/attack
 [CmdletBinding(SupportsShouldProcess = $true, ConfirmImpact = "High")]
 [OutputType([void])]
 param (
+	[Parameter()]
 	[Alias("ComputerName", "CN")]
 	[string] $Domain = [System.Environment]::MachineName,
 
@@ -97,8 +98,8 @@ param (
 #region Initialization
 . $PSScriptRoot\..\..\Config\ProjectSettings.ps1 $PSCmdlet -Domain $Domain
 Write-Debug -Message "[$ThisScript] ParameterSet = $($PSCmdlet.ParameterSetName):$($PSBoundParameters | Out-String)"
-Initialize-Project -Strict
 
+Initialize-Project -Strict
 $Domain = Format-ComputerName $Domain
 
 # User prompt