+++ Test/Ruleset.ProgramInfo/Test-Service.ps1
@@ -79,8 +79,11 @@ if ($Domain -ne [System.Environment]::MachineName)
 	Start-Test "remote dnscache -Session"
 	Test-Service dnscache -Session $SessionInstance
 
-	Start-Test "remote dnscache -Domain"
-	Test-Service dnscache -Domain $Domain -Credential $RemotingCredential
+	if ($false)
+	{
+		Start-Test "remote dnscache -Domain"
+		Test-Service dnscache -Domain $Domain -Credential $RemotingCredential
+	}
 }
 else
 {